Fix: context changes
Made the suggested changes to context

having a slight issue in confess.tst in that the recomended line  setState([ …misdemeanours, {your new data here }]).
when using it in confess.ts i get Type 'Misdemeanour[] | undefined' is not an array type. i tried using? but then eslint complains that is is expecting an expression