feat: test framework
I have attempted to add some tests, however I am running into an issue where they will allways fail because they don't expect the export to be there, but if it isn't the app throws this error and doesn't render:
'confess.test.ts' cannot be compiled under '--isolatedModules' because it is considered a global script file. Add an import, export, or an empty 'export {}' statement to make it a module.