feat: filter
implementing the filter.
Note: there is currently a bug where the filter will block everything until you  change it to something other than none and then back to none, I have attempted to remedy this by updating the filter after calling the list but that doesn't appear to work.
used this approach rather than the more sensible setvisibleList([...visibleList,incident])
because of error: Type 'Misdemeanour[] | undefined' is not an array type. any comments on how to resolve this would be appreciated.