feat: test mock ups
Added some appropriate tests.
added some test id's to misdemeanours.tsx and confess.tsx so that the tests could easily get the data.
added misdemeanour checker to misdemeanours.types so we can check if a misemeanour is valid without the risk of throwing errors.
had to comment out a lot of it due to bizarre errors like not being allowed to import and render not accepting the components.